[patch] migrate to jest and eslint-config
