+++ .eslintrc
@@ -1,28 +1,3 @@
 {
-  "extends": "airbnb-base",
-  "plugins": ["jsdoc"],
-  "env": {
-    "node": true,
-    "mocha": true
-  },
-  "globals": {},
-  "rules": {
-    "id-length": ["error", {"min": 2, "properties": "never", "exceptions": ["_", "i", "j"]}],
-    "import/no-extraneous-dependencies": ["error", {"devDependencies": true}],
-    "indent": ["error", 2, {"SwitchCase": 1}],
-    "no-shadow": ["error", { "allow": ["cb", "err", "done", "resolve", "reject"] }],
-    "jsdoc/check-param-names": "error",
-    "jsdoc/check-tag-names": "error",
-    "jsdoc/check-types": "error",
-    "jsdoc/newline-after-description": "error",
-    "jsdoc/require-description-complete-sentence": "error",
-    "jsdoc/require-hyphen-before-param-description": "error",
-    "jsdoc/require-param-description": "error",
-    "jsdoc/require-param-type": "error",
-    "jsdoc/require-param": "error",
-    "jsdoc/require-returns-description": "error",
-    "jsdoc/require-returns-type": "error",
-    "consistent-return": 2,
-    "max-len": [2, 120]
-  }
+  "extends": "@ahmdigital"
 }
