+++ lib/index.js
@@ -3,21 +3,25 @@ const fs = require('fs');
 const nconf = require('nconf');
 const path = require('path');
 const _ = require('lodash');
-const { match, normalise, validate } = require('./schema');
 
+const { match, normalise, validate } = require('./schema');
 
 exports.make = (originalOpts = {}) => {
   const opts = _.cloneDeep(originalOpts);
   _.defaults(opts, { path: process.env.NODE_CONFIG_DIR }, { path: path.resolve(__dirname, '../../..') });
-  _.defaults(opts, {
-    normalise: true,
-    local: process.env.NODE_CONFIG_FILE,
-    onError: console.error, // eslint-disable-line no-console
-    schema: path.resolve(opts.path, 'schema.json'),
-    default: path.resolve(opts.path, 'default.config.json'),
-  }, {
-    local: path.resolve(opts.path, 'local.config.json'),
-  });
+  _.defaults(
+    opts,
+    {
+      normalise: true,
+      local: process.env.NODE_CONFIG_FILE,
+      onError: console.error, // eslint-disable-line no-console
+      schema: path.resolve(opts.path, 'schema.json'),
+      default: path.resolve(opts.path, 'default.config.json'),
+    },
+    {
+      local: path.resolve(opts.path, 'local.config.json'),
+    },
+  );
 
   const store = new nconf.Provider();
   const env = process.env.NODE_ENV || 'development';
@@ -50,7 +54,9 @@ exports.make = (originalOpts = {}) => {
     .defaults({ NODE_ENV: 'development' });
 
   // Attempt to type-cast string literals to appropriate data types as per schema
-  if (opts.normalise) { normalise(store, schema); }
+  if (opts.normalise) {
+    normalise(store, schema);
+  }
   const errors = validate(store, schema);
   if (errors.length) {
     const { field, message, value, type } = errors[0];
