+++ lib/schema.js
@@ -16,12 +16,13 @@ exports.match = (schema) => {
   };
   walk([], schema);
 
-  const pattern = paths.map(context => context.join('__')).join('|');
+  const pattern = _.invokeMap(paths, 'join', '__').join('|');
   return new RegExp(`^(${pattern})$`);
 };
 
 exports.validateKeys = (data) => {
   const result = [];
+  // eslint-disable-next-line lodash/prefer-lodash-method
   traverse(data).forEach(function context() {
     const path = this.path.join('.');
     if (!validKeyRegex.test(this.key)) {
@@ -47,21 +48,23 @@ exports.validate = (store, originalSchema) => {
 };
 
 const castingFns = {
-  integer: val => parseInt(val, 10),
+  array: (val) => JSON.parse(val),
+  boolean: (val) => val === 'true',
+  integer: (val) => parseInt(val, 10),
   number: Number,
-  boolean: val => (val === 'true'),
-  array: val => JSON.parse(val),
 };
 
 exports.normalise = (store, schema) => {
   const errors = exports.validate(store, schema);
-  if (errors === null) { return; }
+  if (errors === null) {
+    return;
+  }
 
   _(errors)
     .filter({ name: 'type' })
     .filter(({ argument }) => _.has(castingFns, argument[0]))
     .forEach((error) => {
-      const key = error.property.substring(9).replace(/[.]/g, ':'); // convert 'instance.a.b.c' into 'a:b:c'
+      const key = _.replace(error.property.substring(9), /[.]/g, ':'); // convert 'instance.a.b.c' into 'a:b:c'
       const val = castingFns[error.schema.type](store.get(key));
       store.set(key, val);
     });
