+++ package.json
@@ -14,8 +14,8 @@
   },
   "homepage": "https://github.com/ahmdigital/ahm-config",
   "scripts": {
-    "test": "mocha -R spec --recursive",
-    "lint": "eslint -c .eslintrc .",
+    "test": "jest",
+    "lint": "eslint .",
     "release": "scripts/release"
   },
   "dependencies": {
@@ -25,12 +25,9 @@
     "traverse": "^0.6.6"
   },
   "devDependencies": {
-    "eslint": "5.16.0",
-    "eslint-config-airbnb-base": "11.2.0",
-    "eslint-plugin-import": "2.20.2",
-    "eslint-plugin-jsdoc": "24.0.0",
+    "@ahmdigital/eslint-config": "3.1.30",
     "gint": "0.9.2",
-    "mocha": "5.2.0"
+    "jest": "26.0.1"
   },
   "renovate": {
     "extends": [
