+++ test/lo-local.test.js
@@ -1,4 +1,3 @@
-const assert = require('assert');
 const config = require('../lib');
 
 describe('ahm-config: no-local', () => {
@@ -6,11 +5,11 @@ describe('ahm-config: no-local', () => {
 
   it('should work fine when there is no config.local.json', () => {
     const store = config.make({ path });
-    assert.equal(store.get('a'), 1);
+    expect(store.get('a')).toBe(1);
   });
 
   it('should set defaults', () => {
     const store = config.make({ path });
-    assert.equal(store.get('NODE_ENV'), 'development');
+    expect(store.get('NODE_ENV')).toBe('test');
   });
 });
